pyproject.toml
[project]
name = "ocp-heat-reuse-policy-searcher"
version = "1.0.0"
description = "Automated discovery of global data center heat reuse policies"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"

dependencies = [
    "httpx>=0.27.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pdfplumber>=0.10.0",
    "anthropic>=0.39.0",
    "gspread>=6.0.0",
    "google-auth>=2.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
policy-search = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
